+++ new/src/AcademiiSdk/Model/BulkAddUsersToGroupsResponseData.cs
@@ -103,25 +103,25 @@ namespace AcademiiSdk.Model
         {
             var errors = new List<string>();
             // UserCount (long) maximum
-            if (this.UserCount > (long)9007199254740991)
+            if (this.UserCount > 9007199254740991)
             {
                 errors.Add("Invalid value for UserCount, must be a value less than or equal to 9007199254740991.");
             }
 
             // UserCount (long) minimum
-            if (this.UserCount < (long)0)
+            if (this.UserCount < 0)
             {
                 errors.Add("Invalid value for UserCount, must be a value greater than or equal to 0.");
             }
 
             // AddedMembershipCount (long) maximum
-            if (this.AddedMembershipCount > (long)9007199254740991)
+            if (this.AddedMembershipCount > 9007199254740991)
             {
                 errors.Add("Invalid value for AddedMembershipCount, must be a value less than or equal to 9007199254740991.");
             }
 
             // AddedMembershipCount (long) minimum
-            if (this.AddedMembershipCount < (long)0)
+            if (this.AddedMembershipCount < 0)
             {
                 errors.Add("Invalid value for AddedMembershipCount, must be a value greater than or equal to 0.");
             }
