+++ new/src/AcademiiSdk/Model/CourseAnalyticsOverviewResponseInputDataCompletionOverTimeInner.cs
@@ -94,13 +94,13 @@ namespace AcademiiSdk.Model
         {
             var errors = new List<string>();
             // CompletedCount (long) maximum
-            if (this.CompletedCount > (long)9007199254740991)
+            if (this.CompletedCount > 9007199254740991)
             {
                 errors.Add("Invalid value for CompletedCount, must be a value less than or equal to 9007199254740991.");
             }
 
             // CompletedCount (long) minimum
-            if (this.CompletedCount < (long)0)
+            if (this.CompletedCount < 0)
             {
                 errors.Add("Invalid value for CompletedCount, must be a value greater than or equal to 0.");
             }
