+++ new/src/AcademiiSdk/Model/GetVoicesResponseVoicesInnerSettings.cs
@@ -102,37 +102,37 @@ namespace AcademiiSdk.Model
         {
             var errors = new List<string>();
             // Stability (decimal) maximum
-            if (this.Stability > (decimal)1)
+            if (this.Stability > 1)
             {
                 errors.Add("Invalid value for Stability, must be a value less than or equal to 1.");
             }
 
             // Stability (decimal) minimum
-            if (this.Stability < (decimal)0)
+            if (this.Stability < 0)
             {
                 errors.Add("Invalid value for Stability, must be a value greater than or equal to 0.");
             }
 
             // SimilarityBoost (decimal) maximum
-            if (this.SimilarityBoost > (decimal)1)
+            if (this.SimilarityBoost > 1)
             {
                 errors.Add("Invalid value for SimilarityBoost, must be a value less than or equal to 1.");
             }
 
             // SimilarityBoost (decimal) minimum
-            if (this.SimilarityBoost < (decimal)0)
+            if (this.SimilarityBoost < 0)
             {
                 errors.Add("Invalid value for SimilarityBoost, must be a value greater than or equal to 0.");
             }
 
             // Style (decimal) maximum
-            if (this.Style > (decimal)1)
+            if (this.Style > 1)
             {
                 errors.Add("Invalid value for Style, must be a value less than or equal to 1.");
             }
 
             // Style (decimal) minimum
-            if (this.Style < (decimal)0)
+            if (this.Style < 0)
             {
                 errors.Add("Invalid value for Style, must be a value greater than or equal to 0.");
             }
