+++ new/src/AcademiiSdk/Model/UpdateCourseRequest.cs
@@ -166,13 +166,13 @@ namespace AcademiiSdk.Model
             }
 
             // SortOrder (long) maximum
-            if (this.SortOrder > (long)9007199254740991)
+            if (this.SortOrder > 9007199254740991)
             {
                 errors.Add("Invalid value for SortOrder, must be a value less than or equal to 9007199254740991.");
             }
 
             // SortOrder (long) minimum
-            if (this.SortOrder < (long)0)
+            if (this.SortOrder < 0)
             {
                 errors.Add("Invalid value for SortOrder, must be a value greater than or equal to 0.");
             }
